not-found route

+++ src/components/Inventory/Manage/ManageInventory.jsx
@@ -15,12 +15,14 @@ const ManageInventory = () => {
     getFruits();
   }, []);
   const handleDelete = async (id) => {
+    const agree = window.confirm("Are you sure to delete the item?");
+    if (!agree) return;
+
     const res = await axios.delete(`http://localhost:5000/fruits/${id}`);
 
     const deletedFruit = res.data;
     const newFruits = fruits.filter((f) => f._id !== deletedFruit._id);
     setFruits(newFruits);
-    console.log(deletedFruit);
   };
   return (
     <section className="container">
